Move validation schema to util file
