+++ src/utils/validation.js
@@ -1,0 +1,32 @@
+import * as Yup from 'yup';
+
+const INVALID_DATE = 'Invalid date';
+
+// RA Summary: eslint - security/detect-unsafe-regex - Denial of Service: Regular Expression
+// RA: Locates potentially unsafe regular expressions, which may take a very long time to run, blocking the event loop
+// RA: Per MilMove SSP, predisposing conditions are regex patterns from untrusted sources or unbounded matching.
+// RA: The regex pattern is a constant string set at compile-time and it is bounded to 10 characters (zip code).
+// RA Developer Status: Mitigated
+// RA Validator Status:  Mitigated
+// RA Modified Severity: N/A
+// eslint-disable-next-line security/detect-unsafe-regex
+export const ZIP_CODE_REGEX = /^(\d{5}([-]\d{4})?)$/;
+
+// eslint-disable-next-line import/prefer-default-export
+export function validateDate(value) {
+  let error;
+  if (value === INVALID_DATE || !value) {
+    error = 'Required';
+  }
+  return error;
+}
+
+/** Yup validation schemas */
+
+export const requiredAddressSchema = Yup.object().shape({
+  street_address_1: Yup.string().required('Required'),
+  street_address_2: Yup.string(),
+  city: Yup.string().required('Required'),
+  state: Yup.string().length(2, 'Must use state abbreviation').required('Required'),
+  postal_code: Yup.string().matches(ZIP_CODE_REGEX, 'Must be valid zip code').required('Required'),
+});
